refactor: ReportBox, ReportLayout 폴더 경로 변경
