--- src/pages/reports/Report/components/ReportBox/index.ts
@@ -1,3 +1,0 @@
-import ReportBox from "./ReportBox";
-
-export default ReportBox;
