+++ src/pages/reports/Report/components/layout/ReportBox/index.ts
@@ -1,0 +1,3 @@
+import ReportBox from "./ReportBox.tsx";
+
+export default ReportBox;
